Remve int conversion

+++ dns/record/type-srv.go
@@ -19,9 +19,9 @@ func (r *SRV) Encode() []*Answer {
 			Data:  make([]byte, 6+len(tbytes)),
 		}
 
-		binary.BigEndian.PutUint16(answer.Data[0:2], uint16(rec.Priority))
-		binary.BigEndian.PutUint16(answer.Data[2:4], uint16(rec.Weight))
-		binary.BigEndian.PutUint16(answer.Data[4:6], uint16(rec.Port))
+		binary.BigEndian.PutUint16(answer.Data[0:2], rec.Priority)
+		binary.BigEndian.PutUint16(answer.Data[2:4], rec.Weight)
+		binary.BigEndian.PutUint16(answer.Data[4:6], rec.Port)
 		copy(answer.Data[6:], tbytes)
 
 		answers = append(answers, answer)
